requirements.txt
google-api-core==1.31.1
google-api-python-client==2.27.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
gspread==4.0.0
discord-py-interactions==3.0.2
discord.py==1.7.3
firebase-admin==5.0.3